リザルト分岐削除
