+++ Assets/MyGameAssets/Script/Player/ResultPlayerAnimator.cs
@@ -21,9 +21,6 @@ public class ResultPlayerAnimator : SingletonMonoBehaviour<ResultPlayerAnimator>
     [SerializeField]
     Animator playerAnim = default;                      // アニメーター
 
-    [SerializeField]
-    ResultJunction resultJunction = default;            // リザルト分岐クラス
-
     public bool IsEnd { get; private set; } = false;    // アニメーション終了フラグ
 
     /// <summary>
@@ -46,20 +43,7 @@ public class ResultPlayerAnimator : SingletonMonoBehaviour<ResultPlayerAnimator>
         {
             case (int)AnimKind.Title: playerAnim.SetTrigger("Title"); break;
             case (int)AnimKind.Main: playerAnim.SetTrigger("Main"); break;
-            case (int)AnimKind.Result:
-
-                // 開始時にリザルト状態に応じたアニメーションを再生
-                if (!resultJunction.IsJunction)
-                {
-                    // 良い時
-                    playerAnim.SetTrigger("Result");
-                }
-                else
-                {
-                    // 悪い時
-                    playerAnim.SetTrigger("ButResult");
-                }
-                break;
+            case (int)AnimKind.Result:playerAnim.SetTrigger("Result"); break;
 
             case (int)AnimKind.ScoreResult:
